do periodic pricing at approximate interval using platform events
