+++ src/util/common.ts
@@ -1,10 +1,11 @@
 // common utilities and helper functions
 
 import { Address, BigDecimal, BigInt, log } from '@graphprotocol/graph-ts';
-
-import { Platform, User, PoolDayData, Pool } from '../../generated/schema'
-import { ZERO_BIG_INT, ZERO_BIG_DECIMAL, ZERO_ADDRESS } from '../util/constants'
-
+import { Platform, User, PoolDayData, Pool, Token } from '../../generated/schema'
+import { ZERO_BIG_INT, ZERO_BIG_DECIMAL, ZERO_ADDRESS, PRICING_PERIOD } from '../util/constants'
+import { GeyserV1 as GeyserContractV1 } from '../../generated/templates/GeyserV1/GeyserV1'
+import { updateGeyserV1 } from '../util/geyserv1'
+import { updatePool } from '../util/pool'
 
 export function integerToDecimal(value: BigInt, decimals: BigInt = BigInt.fromI32(18)): BigDecimal {
   let denom = BigInt.fromI32(10).pow(decimals.toI32() as u8);
@@ -28,6 +29,7 @@ export function createNewPlatform(): Platform {
   platform.gysrSpent = ZERO_BIG_DECIMAL;
   platform.volume = ZERO_BIG_DECIMAL;
   platform._activePools = [];
+  platform._updated = ZERO_BIG_INT;
 
   return platform;
 }
@@ -53,4 +55,56 @@ export function updatePoolDayData(pool: Pool, timestamp: number): PoolDayData {
   poolDayData.usage = pool.usage;
 
   return poolDayData!;
+}
+
+
+export function updatePlatform(platform: Platform, timestamp: BigInt, skip: Pool): boolean {
+  // skip if pricing period has not elapsed
+  if (timestamp.minus(platform._updated).lt(PRICING_PERIOD)) {
+    return false;
+  }
+
+  let pools = platform._activePools;
+  var stale: string[] = [];
+
+  for (let i = 0; i < pools.length; i++) {
+    // don't need to price pool that triggered this event
+    if (pools[i] == skip.id) {
+      continue;
+    }
+
+    // load
+    let pool = Pool.load(pools[i])!;
+    let stakingToken = Token.load(pool.stakingToken)!;
+    let rewardToken = Token.load(pool.rewardToken)!;
+
+    // update pool
+    if (pool.poolType == 'GeyserV1') {
+      let contract = GeyserContractV1.bind(Address.fromString(pool.id));
+      updateGeyserV1(pool, platform!, contract, stakingToken, rewardToken, timestamp);
+    } else {
+      updatePool(pool, platform!, stakingToken, rewardToken, timestamp);
+    }
+
+    // update pool day snapshot
+    let poolDayData = updatePoolDayData(pool, timestamp.toI32());
+
+    // store
+    pool.save();
+    stakingToken.save();
+    rewardToken.save();
+    poolDayData.save();
+
+    // remove from priced pool list if stale
+    if (pool.state == 'Stale') {
+      stale.push(pool.id);
+      log.info('Marking pool as stale {}', [pool.id.toString()]);
+    }
+  }
+
+  if (stale.length) {
+    platform._activePools = pools.filter((x) => !stale.includes(x));
+  }
+  platform._updated = timestamp;
+  return true;
 }
